crate2nix workflow: update to use HEAD of crate2nix
crate2nix 0.8.0 has a bug relating to renamed crates where it
fails to correctly identify all the dependencies - this leads
to one of the versions of tokio that gets built in failing to
get the right dependencies, which causes it to actually fail
to build.

Fix this by pinning to the current HEAD version for the
moment.